curl or wget will work both
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@5606 8aca7d54-2c30-db11-9de9-000461428c89

+++ get_lang_files.sh
@@ -6,15 +6,36 @@ cd simutrans/text
 # get the translations for basis
 # The first file is longer, but only because it contains SQL error messages
 # - discard it after complete download (although parsing it would give us the archive's name):
-wget --post-data "version=0&choice=all&submit=Export!"  --delete-after http://simutrans-germany.com/translator/script/main.php?page=wrap || {
-  echo "Error: generating file language_pack-Base+texts.zip failed (wget returned $?)" >&2;
-  exit 3;
-}
-wget -N http://simutrans-germany.com/translator/data/tab/language_pack-Base+texts.zip || {
-  echo "Error: download of file language_pack-Base+texts.zip failed (wget returned $?)" >&2
-  rm -f "language_pack-Base+texts.zip"
-  exit 4
-}
+
+# Use curl if available, else use wget
+curl -h > /dev/null
+if [ $? -eq 0 ]; then
+    curl -d "version=0&choice=all&submit=Export%21" http://simutrans-germany.com/translator/script/main.php?page=wrap > /dev/null || {
+      echo "Error: generating file language_pack-Base+texts.zip failed (curl returned $?)" >&2;
+      exit 3;
+    }
+    curl http://simutrans-germany.com/translator/data/tab/language_pack-Base+texts.zip > language_pack-Base+texts.zip || {
+      echo "Error: download of file language_pack-Base+texts.zip failed (curl returned $?)" >&2
+      rm -f "language_pack-Base+texts.zip"
+      exit 4
+    }
+else
+    wget --help > /dev/null
+    if [ $? -eq 0 ]; then
+        wget --post-data "version=0&choice=all&submit=Export!"  --delete-after http://simutrans-germany.com/translator/script/main.php?page=wrap || {
+          echo "Error: generating file language_pack-Base+texts.zip failed (wget returned $?)" >&2;
+          exit 3;
+        }
+        wget -N http://simutrans-germany.com/translator/data/tab/language_pack-Base+texts.zip || {
+          echo "Error: download of file language_pack-Base+texts.zip failed (wget returned $?)" >&2
+          rm -f "language_pack-Base+texts.zip"
+          exit 4
+        }
+    else
+        echo "Error: Neither curl or wget are available on your system, please install either and try again!" >&2
+        exit 6
+    fi
+fi
 unzip -tv "language_pack-Base+texts.zip" || {
    echo "Error: file language_pack-Base+texts.zip seems to be defective" >&2
    rm -f "language_pack-Base+texts.zip"
